Move "dagger" test case for Ethash suite
